feature(players): players list updating in real time when players added
