+++ client/src/components/StartGameButton.jsx
@@ -1,0 +1,15 @@
+// LIBRARY IMPORTS
+import { useEffect } from 'react'
+import axios from 'axios';
+
+export default function StartGameButton({ lobby, player, playerList, socket }) {
+  //HANDLE EVENTS
+  const handleStartGameClick = () => {
+
+  }
+  return (
+    <button>
+      Start Game
+    </button>
+  )
+};
