+++ server/routes.js
@@ -4,7 +4,11 @@ import express from 'express';
 const router = express.Router();
 
 // LOCAL IMPORTS
-import { createLobby, getAllLobbies, getLobbyPlayers, joinLobby, leaveLobby } from './controllers.js';
+import { createLobby } from './controllers/createLobby.js';
+import { getAllLobbies } from './controllers/getAllLobbies.js';
+import { getLobbyPlayers } from './controllers/getLobbyPlayers.js';
+import { joinLobby } from './controllers/joinLobby.js';
+import { leaveLobby } from './controllers/leaveLobby.js';
 
 // ROUTES
 router.post('/lobbies', createLobby);
